JBPM-6841: Review details page layout. Details alignment, input fields size, primary button review and task context tab removed
