+++ jbpm-wb-executor-service/jbpm-wb-executor-service-client/src/main/java/org/jbpm/workbench/es/client/editors/errordetails/ExecutionErrorBasicDetailsViewImpl.html
@@ -4,95 +4,95 @@
             <div class="form-horizontal">
 
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Id"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Id"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-errorId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Type"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Type"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-type"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="DeploymentId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="DeploymentId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-deploymentId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Process_Instance_Id"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Process_Instance_Id"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-processInstanceId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="ProcessId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="ProcessId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-processId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="ActivityId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="ActivityId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-activityId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="ActivityName"></label>
+                    <label class="control-label col-md-2" data-i18n-key="ActivityName"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-activityName"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="JobId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="JobId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-jobId"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Message"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Message"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-errorMessage"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Acknowledged"></label>
-                    <div class="col-md-8">
+                    <label class="control-label col-md-2" data-i18n-key="Acknowledged"></label>
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-acknowledged"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="AcknowledgedBy"></label>
-                    <div class="col-md-8">
+                    <label class="control-label col-md-2" data-i18n-key="AcknowledgedBy"></label>
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-acknowledgedBy"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="AcknowledgedAt"></label>
-                    <div class="col-md-8">
+                    <label class="control-label col-md-2" data-i18n-key="AcknowledgedAt"></label>
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-acknowledgedAt"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="ErrorDate"></label>
-                    <div class="col-md-8">
+                    <label class="control-label col-md-2" data-i18n-key="ErrorDate"></label>
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="error-errorDate"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Error"></label>
-                    <div class="col-md-8">
+                    <label class="control-label col-md-2" data-i18n-key="Error"></label>
+                    <div class="col-md-10">
                         <textarea class="col-sm-12" rows="10" readonly data-field="error-error"></textarea>
                     </div>
                 </div>
