+++ jbpm-wb-executor-service/jbpm-wb-executor-service-client/src/main/java/org/jbpm/workbench/es/client/editors/errordetails/ExecutionErrorDetailsViewImpl.html
@@ -3,7 +3,7 @@
     <div>
         <!-- navigation tabs -->
         <ul class="nav nav-tabs nav-tabs-pf" id="navTabs">
-            <li class="active"><a href="#" data-i18n-key="ErrorDetails"></a></li>
+            <li class="active"><a href="#" data-i18n-key="Details"></a></li>
         </ul>
     </div>
