+++ jbpm-wb-executor-service/jbpm-wb-executor-service-client/src/main/java/org/jbpm/workbench/es/client/editors/jobdetails/JobBasicDetailsViewImpl.html
@@ -3,67 +3,67 @@
         <div class="col-md-12">
             <div class="form-horizontal">
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="JobId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="JobId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-jobId"></span>
                     </div>
                 </div>
                 <div data-field="businesskey_div">
                     <div class="form-group">
-                        <label class="control-label col-md-4" data-i18n-key="BusinessKey"></label>
+                        <label class="control-label col-md-2" data-i18n-key="BusinessKey"></label>
 
-                        <div class="col-md-8">
+                        <div class="col-md-10">
                             <span class="form-control-static" data-field="job-key"></span>
                         </div>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Status"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Status"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-status"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Due_On"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Due_On"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-time"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="CommandName"></label>
+                    <label class="control-label col-md-2" data-i18n-key="CommandName"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-commandName"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Message"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Message"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-message"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Retries"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Retries"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-retries"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" data-i18n-key="Executions"></label>
+                    <label class="control-label col-md-2" data-i18n-key="Executions"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-executions"></span>
                     </div>
                 </div>
                 <div class="form-group" data-field="deployment_id_div">
-                    <label class="control-label col-md-4" data-i18n-key="DeploymentId"></label>
+                    <label class="control-label col-md-2" data-i18n-key="DeploymentId"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <span class="form-control-static" data-field="job-deploymentId"></span>
                     </div>
                 </div>
