+++ jbpm-wb-executor-service/jbpm-wb-executor-service-client/src/main/java/org/jbpm/workbench/es/client/editors/jobdetails/JobDetailsViewImpl.html
@@ -3,7 +3,7 @@
     <div>
         <!-- navigation tabs -->
         <ul class="nav nav-tabs nav-tabs-pf" id="navTabs">
-            <li class="active"><a href="#" data-i18n-key="Job_Details"></a></li>
+            <li class="active"><a href="#" data-i18n-key="Details"></a></li>
         </ul>
     </div>
 
@@ -14,15 +14,15 @@
             <div class="col-md-12">
                 <div class="form-horizontal">
                     <div class="form-group" data-field="errorControlGroup">
-                        <label class="control-label col-md-4" data-i18n-key="Exceptions_Occurred"></label>
-                        <div class="col-md-8">
+                        <label class="control-label col-md-2" data-i18n-key="Exceptions_Occurred"></label>
+                        <div class="col-md-10">
                             <textarea class="col-sm-12"
                                       rows="10" readonly data-field="job-errors"></textarea>
                         </div>
                     </div>
                     <div class="form-group hidden" data-field="params-form-group">
-                        <label class="control-label col-md-4" data-i18n-key="Execution_Parameters"></label>
-                        <div class="col-md-8">
+                        <label class="control-label col-md-2" data-i18n-key="Execution_Parameters"></label>
+                        <div class="col-md-10">
                             <table class="table table-striped table-bordered table-hover">
                                 <thead>
                                 <tr>
