+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/editors/taskadmin/TaskAdminViewImpl.html
@@ -5,23 +5,23 @@
                 <fieldset>
                     <legend data-field="adminDetailsAccordionLabel"></legend>
                     <div class="form-group">
-                        <label class="control-label col-md-3" for="adminUsersGroupsControlsPanel" data-field="adminUsersGroupsControlsLabel"></label>
+                        <label class="control-label col-md-2" for="adminUsersGroupsControlsPanel" data-field="adminUsersGroupsControlsLabel"></label>
 
-                        <div class="col-md-9">
+                        <div class="col-md-10">
                             <p class="form-control-static" id="adminUsersGroupsControlsPanel" data-field="adminUsersGroupsControlsPanel" />
                         </div>
                     </div>
                     <div class="form-group">
-                        <label class="control-label col-md-3" for="adminUserOrGroupText" data-field="adminUserOrGroupLabel"></label>
+                        <label class="control-label col-md-2" for="adminUserOrGroupText" data-field="adminUserOrGroupLabel"></label>
 
-                        <div class="col-md-9">
+                        <div class="col-md-2">
                             <input type="text" data-field="adminUserOrGroupText" id="adminUserOrGroupText" placeholder="" class="form-control"/>
                         </div>
                     </div>
 
                     <div class="form-group">
-                        <div class="col-md-offset-3 col-md-9">
-                            <button class="btn btn-default" data-field="adminForwardButton" id="adminForwardButton"></button>
+                        <div class="col-md-offset-2 col-md-10 ">
+                            <button class="btn btn-primary" data-field="adminForwardButton" id="adminForwardButton"></button>
                         </div>
                     </div>
 
@@ -31,15 +31,15 @@
                 <fieldset>
                     <legend data-field="reminderDetailsAccordionLabel"></legend>
                     <div class="form-group">
-                        <label class="control-label col-md-3" for="actualOwnerPanel" data-field="actualOwnerLabel"></label>
+                        <label class="control-label col-md-2" for="actualOwnerPanel" data-field="actualOwnerLabel"></label>
 
-                        <div class="col-md-9">
+                        <div class="col-md-10">
                             <p class="form-control-static" id="actualOwnerPanel" data-field="actualOwnerPanel" />
                         </div>
                     </div>
                     <div class="form-group">
-                        <div class="col-md-offset-3 col-md-9">
-                            <button class="btn btn-default" data-field="adminReminderButton" id="adminReminderButton"></button>
+                        <div class="col-md-offset-2 col-md-10">
+                            <button class="btn btn-primary" data-field="adminReminderButton" id="adminReminderButton"></button>
                         </div>
                     </div>
                 </fieldset>
