+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/editors/taskassignments/TaskAssignmentsViewImpl.html
@@ -3,23 +3,23 @@
         <div class="form-horizontal">
             <fieldset>
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="usersGroupsControlsPanel" data-field="usersGroupsControlsLabel"></label>
+                    <label class="control-label col-md-2" for="usersGroupsControlsPanel" data-field="usersGroupsControlsLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-10">
                         <ul id="usersGroupsControlsPanel" data-field="usersGroupsControlsPanel" class="list-unstyled"></ul>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="userOrGroupText" data-field="userOrGroupLabel"></label>
+                    <label class="control-label col-md-2" for="userOrGroupText" data-field="userOrGroupLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-2">
                         <input type="text" data-field="userOrGroupText" id="userOrGroupText" placeholder="" class="form-control"/>
                         <span id="helpBlock" class="help-block" for="userOrGroupText" data-field="userOrGroupHelpBlock"></span>
                     </div>
                 </div>
                 <div class="form-group">
-                    <div class="col-md-offset-3 col-md-9">
-                        <button class="btn btn-default" data-field="delegateButton" id="delegateButton"></button>
+                    <div class="col-md-offset-2 col-md-10">
+                        <button class="btn btn-primary" data-field="delegateButton" id="delegateButton"></button>
                     </div>
                 </div>
             </fieldset>
