+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/editors/taskcomments/TaskCommentsViewImpl.html
@@ -3,23 +3,22 @@
         <div class="form-horizontal">
             <fieldset>
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="newTaskCommentTextArea" data-field="newTaskCommentLabel"></label>
-                    <div class="col-md-9">
+                    <label class="control-label col-md-2" for="newTaskCommentTextArea" data-field="newTaskCommentLabel"></label>
+                    <div class="col-md-5">
                         <input type="text" data-field="newTaskCommentTextArea" id="newTaskCommentTextArea" placeholder="" class="form-control"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <div class="col-md-offset-3 col-md-9">
+                    <div class="col-md-offset-2 col-md-10">
                         <button class="btn btn-primary" data-field="addCommentButton" id="addCommentButton"></button>
                     </div>
                 </div>
             </fieldset>
         </div>
         <div data-field="listContainer">
-            <div id="commentsListGrid" data-field="commentsListGrid" />
-            <!--<table id="commentsListGrid" data-field="commentsListGrid" class="table table-bordered table-striped table-hover">-->
-            <!--</table>-->
-            <!--<div id="pager" class="pagination pagination-right pull-right" data-field="pager"></div>-->
+            <div class="col-md-12">
+                <div id="commentsListGrid" data-field="commentsListGrid"/>
+            </div>
         </div>
     </div>
 </div>
