+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/editors/taskdetails/TaskDetailsViewImpl.html
@@ -3,44 +3,59 @@
         <div class="form-horizontal">
             <fieldset>
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="taskDescriptionTextArea" data-field="taskDescriptionLabel"></label>
+                    <label class="control-label col-md-2" for="taskDescriptionTextArea" data-field="taskDescriptionLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-5">
                         <input type="text" data-field="taskDescriptionTextArea" id="taskDescriptionTextArea" placeholder="" class="form-control"/>
                     </div>
                 </div>
 
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="taskStatusText" data-field="taskStatusLabel"></label>
+                    <label class="control-label col-md-2" for="taskStatusText" data-field="taskStatusLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-10">
                         <p data-field="taskStatusText" id="taskStatusText" class="form-control-static"/>
                     </div>
                 </div>
 
 
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="dueDate" data-field="dueDateLabel"></label>
+                    <label class="control-label col-md-2" for="dueDate" data-field="dueDateLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-2">
                         <div data-field="dueDate" id="dueDate" class="col-md-8"></div>
                         <div data-field="dueDateTime" id="dueDateTime" class="col-md-4"></div>
                     </div>
                 </div>
 
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="taskPriorityListBox" data-field="taskPriorityLabel"></label>
+                    <label class="control-label col-md-2" for="taskPriorityListBox" data-field="taskPriorityLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-2">
                         <select data-field="taskPriorityListBox" id="taskPriorityListBox" class="form-control"></select>
                     </div>
                 </div>
 
                 <div class="form-group">
-                    <label class="control-label col-md-3" for="userText" data-field="userLabel"></label>
+                    <label class="control-label col-md-2" for="userText" data-field="userLabel"></label>
 
-                    <div class="col-md-9">
+                    <div class="col-md-10">
                         <p data-field="userText" id="userText" class="form-control-static"/>
+                    </div>
+                </div>
+                <div class="form-group">
+                    <label class="control-label col-md-2" for="processInstanceIdText" data-field="processInstanceIdLabel"></label>
+
+                    <div class="col-md-10">
+                        <p data-field="processInstanceIdText" id="processInstanceIdText" class="form-control-static"/>
+                    </div>
+                </div>
+
+                <div class="form-group">
+                    <label class="control-label col-md-2" for="processIdText" data-field="processIdLabel"></label>
+
+                    <div class="col-md-10">
+                        <p data-field="processIdText" id="processIdText" class="form-control-static"/>
                     </div>
                 </div>
             </fieldset>
@@ -48,8 +63,8 @@
         <div class="form-horizontal">
             <fieldset>
                 <div class="form-group">
-                    <div class="col-md-offset-3 col-md-9">
-                        <button class="btn btn-default" data-field="updateTaskButton" id="updateTaskButton"></button>
+                    <div class="col-md-offset-2 col-md-10">
+                        <button class="btn btn-primary" data-field="updateTaskButton" id="updateTaskButton"></button>
                     </div>
                 </div>
             </fieldset>
