--- jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/editors/taskprocesscontext/TaskProcessContextViewImpl.html
@@ -1,29 +1,0 @@
-<div class="row" style="padding-top: 10px">
-    <div class="col-md-12">
-        <div class="form-horizontal">
-            <fieldset>
-                <div class="form-group">
-                    <label class="control-label col-md-3" for="processInstanceIdText" data-field="processInstanceIdLabel"></label>
-
-                    <div class="col-md-9">
-                        <p data-field="processInstanceIdText" id="processInstanceIdText" class="form-control-static"/>
-                    </div>
-                </div>
-
-                <div class="form-group">
-                    <label class="control-label col-md-3" for="processIdText" data-field="processIdLabel"></label>
-
-                    <div class="col-md-9">
-                        <p data-field="processIdText" id="processIdText" class="form-control-static"/>
-                    </div>
-                </div>
-
-                <div class="form-group">
-                    <div class="col-md-offset-3 col-md-9">
-                        <button class="btn btn-default" data-field="pIDetailsButton" id="pIDetailsButton"></button>
-                    </div>
-                </div>
-            </fieldset>
-        </div>
-    </div>
-</div>
