+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/diagram/ProcessDiagramWidgetView.html
@@ -14,5 +14,5 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   -->
-<div class="container-fluid" id="processDiagramDiv" style="overflow:auto; height:600px">
+<div class="container-fluid" id="processDiagramDiv" style="overflow:auto; padding-top: 10px;">
 </div>
