+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/instance/details/ProcessInstanceDetailsViewImpl.html
@@ -4,58 +4,58 @@
             <div class="form-horizontal">
 
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="processDefinitionIdText" data-field="processDefinitionIdLabel"></label>
+                    <label class="control-label col-md-2" for="processDefinitionIdText" data-field="processDefinitionIdLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="processDefinitionIdText" id="processDefinitionIdText"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="stateText" data-field="stateLabel"></label>
+                    <label class="control-label col-md-2" for="stateText" data-field="stateLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="stateText" id="stateText"/>
                     </div>
                 </div>
                 <div class="form-group ">
-                    <label class="control-label col-md-4" for="processDeploymentText" data-field="processDeploymentLabel"></label>
+                    <label class="control-label col-md-2" for="processDeploymentText" data-field="processDeploymentLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="processDeploymentText" id="processDeploymentText"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="processVersionText" data-field="processVersionLabel"></label>
+                    <label class="control-label col-md-2" for="processVersionText" data-field="processVersionLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="processVersionText" id="processVersionText"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="correlationKeyText" data-field="correlationKeyLabel"></label>
+                    <label class="control-label col-md-2" for="correlationKeyText" data-field="correlationKeyLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="correlationKeyText" id="correlationKeyText"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="parentProcessInstanceIdText" data-field="parentProcessInstanceIdLabel"></label>
+                    <label class="control-label col-md-2" for="parentProcessInstanceIdText" data-field="parentProcessInstanceIdLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="parentProcessInstanceIdText" id="parentProcessInstanceIdText"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="activeTasksListBox" data-field="activeTasksListLabel"></label>
+                    <label class="control-label col-md-2" for="activeTasksListBox" data-field="activeTasksListLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="activeTasksListBox" id="activeTasksListBox"/>
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-4" for="currentActivitiesListBox" data-field="currentActivitiesListLabel"></label>
+                    <label class="control-label col-md-2" for="currentActivitiesListBox" data-field="currentActivitiesListLabel"></label>
 
-                    <div class="col-md-8">
+                    <div class="col-md-10">
                         <p class="form-control-static" data-field="currentActivitiesListBox" id="currentActivitiesListBox"/>
                     </div>
                 </div>
